Enable GitHub Actions for Testing
Wire up an initial GitHib Actions flow for testing only. This breaks
Travis but it was already broken due to bit-rot.

This moves to a more modern base OS of Ubuntu 20.04. With this we can
now depend of a recent version of hunspell and associated dictionaries.
This dramatically reduces CI time and repo size.